requirements.txt
playwright==1.52.0
openai==1.108.1
pydantic==2.9.2
python-dotenv==1.1.1
requests==2.32.3
pytest==8.4.2
mypy==1.19.1
pillow==10.4.0
